# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /storage/application-uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestApplicationUploadUrl
      summary: Request a presigned URL for a public application-form upload
      description: |
        Public (no auth). Used by prospective students to upload their completed
        S1/S5 application form. Rate-limited per IP and restricted to common
        document/image types. Returns a presigned GCS URL for direct upload.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: Too many uploads from this IP.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      description: |
        Serves object entities uploaded via presigned URLs. These can optionally
        be protected with authentication or ACL checks based on the use case.
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
          description: Object path within the private object dir (e.g. `uploads/some-uuid`).
      responses:
        '200':
          description: Object content streamed with correct Content-Type.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to serve object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /resources:
    get:
      tags: [Resources]
      operationId: listResources
      summary: List student resources
      description: Returns all past papers and holiday work, newest first.
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
          description: Optional filter by category (past_paper or holiday_work).
      responses:
        '200':
          description: List of resources.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceList'
    post:
      tags: [Resources]
      operationId: createResource
      summary: Create a resource record
      description: |
        Records metadata for an uploaded file. The file must already be uploaded
        via the presigned URL flow; pass the returned objectPath here.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResourceInput'
      responses:
        '201':
          description: Resource created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /resources/{id}:
    patch:
      tags: [Resources]
      operationId: updateResource
      summary: Update a resource
      description: |
        Update resource metadata and optionally replace the file (pass a new
        objectPath from the upload flow). Broadcasts so the public site refreshes.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Resource id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateResourceInput'
      responses:
        '200':
          description: Resource updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        '400':
          description: Missing or invalid fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [Resources]
      operationId: deleteResource
      summary: Delete a resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Resource id.
      responses:
        '204':
          description: Resource deleted.
        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  schemas:
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
          description: Original file name.
        size:
          type: integer
          minimum: 1
          description: File size in bytes.
        contentType:
          type: string
          minLength: 1
          description: MIME type of the file (e.g. `image/jpeg`).
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
          description: Presigned GCS URL for PUT upload.
        objectPath:
          type: string
          description: Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database.
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'
    Resource:
      type: object
      required: [id, title, subject, category, level, objectPath, fileName, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        subject:
          type: string
        category:
          type: string
          description: past_paper or holiday_work.
        level:
          type: string
          description: 'O-Level, A-Level, or All.'
        term:
          type: string
          nullable: true
        objectPath:
          type: string
        fileName:
          type: string
        fileSize:
          type: integer
          nullable: true
        contentType:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    ResourceList:
      type: array
      items:
        $ref: '#/components/schemas/Resource'
    CreateResourceInput:
      type: object
      required: [title, subject, category, objectPath, fileName]
      properties:
        title:
          type: string
          minLength: 1
        subject:
          type: string
          minLength: 1
        category:
          type: string
          minLength: 1
          description: past_paper or holiday_work.
        level:
          type: string
          description: 'O-Level, A-Level, or All. Defaults to All.'
        term:
          type: string
          nullable: true
        objectPath:
          type: string
          minLength: 1
        fileName:
          type: string
          minLength: 1
        fileSize:
          type: integer
          nullable: true
        contentType:
          type: string
          nullable: true
    UpdateResourceInput:
      type: object
      description: Partial update. Omit objectPath to keep the existing file.
      properties:
        title:
          type: string
          minLength: 1
        subject:
          type: string
          minLength: 1
        category:
          type: string
          minLength: 1
          description: past_paper, holiday_work, or application_form.
        level:
          type: string
        term:
          type: string
          nullable: true
        objectPath:
          type: string
          minLength: 1
        fileName:
          type: string
          minLength: 1
        fileSize:
          type: integer
          nullable: true
        contentType:
          type: string
          nullable: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...