# lib/api-zod/src/generated/types/updateResourceInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Partial update. Omit objectPath to keep the existing file.
 */
export interface UpdateResourceInput {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  subject?: string;
  /**
     * past_paper, holiday_work, or application_form.
     * @minLength 1
     */
  category?: string;
  level?: string;
  term?: string | null;
  /** @minLength 1 */
  objectPath?: string;
  /** @minLength 1 */
  fileName?: string;
  fileSize?: number | null;
  contentType?: string | null;
}